# services/auth-policy/src/main/resources/openapi/gcs-saker-operations.openapi.yaml
openapi: 3.1.0
info:
  title: GCS-Saker Operations API
  version: "2026-08-20"
  description: |
    운영 배포의 public edge 기준 통합 계약이다. 내부 포트, DB, Redis, MQTT 관리 포트와
    credential 원문은 문서화하거나 노출하지 않는다. Swagger UI는 조회 전용이며 실제 변경
    요청은 승인된 운영 절차와 전용 클라이언트에서 수행한다.
servers:
  - url: https://a4ai.121-159-26-245.sslip.io
    description: Production edge
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/groups:
    get: { tags: [Groups], summary: principal이 조회 가능한 그룹, parameters: [ { $ref: '#/components/parameters/PageLimit' }, { $ref: '#/components/parameters/PageOffset' } ], responses: { '200': { $ref: '#/components/responses/Json' } } }
  /api/v1/groups/{groupId}/devices:
    get: { tags: [Groups], summary: 그룹 장비 조회, parameters: [ { $ref: '#/components/parameters/GroupId' }, { $ref: '#/components/parameters/PageLimit' }, { $ref: '#/components/parameters/PageOffset' } ], responses: { '200': { $ref: '#/components/responses/Json' }, '403': { $ref: '#/components/responses/Error' } } }
  /api/v1/groups/{groupId}/dashboard:
    get: { tags: [Groups], summary: 그룹 dashboard aggregate, parameters: [ { $ref: '#/components/parameters/GroupId' } ], responses: { '200': { $ref: '#/components/responses/Json' } } }
...
  /media-control/api/v1/streams/{streamId}:
    get: { tags: [Media], summary: stream 상세, parameters: [ { $ref: '#/components/parameters/StreamId' } ], responses: { '200': { $ref: '#/components/responses/Json' } } }
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
    csrfCookie: { type: apiKey, in: header, name: X-GCS-CSRF, description: HttpOnly refresh cookie와 함께 same-origin 값을 사용 }
    deviceUuid: { type: apiKey, in: header, name: X-GCS-Device-UUID }
    deviceCredential: { type: apiKey, in: header, name: X-GCS-Device-Credential, description: 원문을 로그·문서·브라우저 저장소에 기록하지 않는다 }
  parameters:
...
    GroupId: { name: groupId, in: path, required: true, schema: { type: string } }
    StreamId: { name: streamId, in: path, required: true, schema: { type: string } }
...
    PageLimit: { name: limit, in: query, required: false, schema: { type: integer, minimum: 1, maximum: 500, default: 200 } }
    PageOffset: { name: offset, in: query, required: false, schema: { type: integer, minimum: 0, maximum: 100000, default: 0 } }
...
  responses:
    Json: { description: Successful JSON response, content: { application/json: { schema: { $ref: '#/components/schemas/JsonObject' } } } }
    Error: { description: Stable external error without stack trace or secret, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
...
  schemas:
    JsonObject: { type: object, additionalProperties: true }
...
    Error: { type: object, additionalProperties: false, properties: { detail: { type: string }, code: { type: string }, traceId: { type: string } } }
...